Removed hidden div and instead passed variables in functions, refactored settingscard.py to two ddifferent files
