+++ .gitignore
@@ -144,3 +144,4 @@ ukbb_data_field_hierarchy.csv
 ukbb_meta_data_subset.csv
 ukbb_meta_data_21079.csv
 /ukbb-data-explorer/
+node_modules
