+++ src/app.py
@@ -5,6 +5,7 @@
 import os
 import sys
 import dash_html_components as html
+import dash_core_components as dcc
 import dash_bootstrap_components as dbc
 from dash.dependencies import Input, Output, State
